chore: decouple saving state from tuner
