+++ finetuner/tuner/callback/__init__.py
@@ -2,4 +2,4 @@ from .base import BaseCallback  # noqa: F401
 from .progress_bar import ProgressBarCallback  # noqa: F401
 from .wandb_logger import WandBLogger  # noqa: F401
 from .best_model_checkpoint import BestModelCheckpoint
-from .training_model_checkpoint import TrainingCheckpoint
+from .training_checkpoint import TrainingCheckpoint
